+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.0.0-beta2] - In progress
+
+### Added
+
+- New method `Query::map(): array` to apply function to all results of query and get array result
+
 ## [2.0.0-beta1] - 2021-05-11
 
 ### Added